feat: 배너 / 인트로 이미지 컴포넌트화
