+++ src/app/profile/my-activities/post/_components/intro-image-uploader.tsx
@@ -1,0 +1,73 @@
+import Image from "next/image";
+
+interface IntroImagesUploaderProps {
+  introImages: string[];
+  setIntroImages: React.Dispatch<React.SetStateAction<string[]>>;
+}
+
+export default function IntroImagesUploader({
+  introImages,
+  setIntroImages,
+}: IntroImagesUploaderProps) {
+  const handleIntroUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
+    if (event.target.files) {
+      const files = Array.from(event.target.files)
+        .map((file) => URL.createObjectURL(file))
+        .slice(0, 4); // 최대 4개 제한
+      setIntroImages((prev) => [...prev, ...files].slice(0, 4));
+    }
+  };
+
+  const removeIntroImage = (index: number) => {
+    setIntroImages((prev) => prev.filter((_, i) => i !== index));
+  };
+
+  return (
+    <div className="flex flex-col gap-[2.4rem]">
+      <h3 className="text-2xl font-bold">소개 이미지</h3>
+      <div className="flex flex-row gap-4 flex-wrap items-center">
+        {introImages.map((img, index) => (
+          <div
+            key={index}
+            className="relative w-[18rem] h-[18rem] border border-gray-300 rounded-md"
+          >
+            <Image
+              src={img}
+              alt="소개 이미지"
+              layout="fill"
+              objectFit="cover"
+              className="rounded-md"
+            />
+            <button
+              onClick={() => removeIntroImage(index)}
+              className="absolute top-0 right-0 bg-gray-700 text-white text-xs px-2 py-1 rounded-full"
+            >
+              X
+            </button>
+          </div>
+        ))}
+
+        {introImages.length < 4 && (
+          <label className="cursor-pointer w-[18rem] h-[18rem] border border-gray-300 rounded-md flex items-center justify-center">
+            <input
+              type="file"
+              accept="image/*"
+              multiple
+              onChange={handleIntroUpload}
+              className="hidden"
+            />
+            <Image
+              src="/image/set_image_btn.svg"
+              alt="이미지 등록"
+              width={180}
+              height={180}
+            />
+          </label>
+        )}
+      </div>
+      <p className="text-sm text-gray-500">
+        *이미지는 최대 4개까지 등록 가능합니다.
+      </p>
+    </div>
+  );
+}
